chore: move hh service and rename it
